requirements.txt
streamlit>=1.35.0
opencv-contrib-python-headless
numpy
Pillow
streamlit-webrtc